package rename `internal` -> `types` to be exposed
