+++ types/listen_message_event.go
@@ -1,4 +1,4 @@
-package internal
+package types
 
 import (
 	"context"
