+++ types/shared_worker_global_scope.go
@@ -1,6 +1,6 @@
 //go:build js && wasm
 
-package internal
+package types
 
 import (
 	"context"
